Pair internal/auth test files with their source files
